Move `GetPushRulesResponse` to internal package
